requirements.txt
fastapi
uvicorn
langchain
langserve
langchain-openai
requests
python-dotenv